pyproject.toml
[tool.poetry]
name = "attention_control"
version = "0.1.0"
description = ""
authors = ["Viktor Vlasov"]
readme = "README.md"
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.0.0"
flake8 = "^6.0.0"
tqdm = "^4.65.0"
pandas = "^2.0.1"
numpy = "^1.24.3"
pillow = "^9.5.0"
matplotlib = "^3.7.1"
scikit-learn = "^1.2.2"
lightning = "^2.0.2"
efficientnet-pytorch = "^0.7.1"
torchvision = "^0.15.1"
albumentations = "^1.3.0"
uplink = "^0.9.7"
yacs = "^0.1.8"
dvc = {extras = ["s3"], version = "^2.55.0"}
ruff = "^0.0.265"
bandit = "^1.7.5"
split-folders = "^0.5.1"
mlflow = "^2.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
skips = ["B101"]